Rename : 폴더명 interfaces -> types로 변경
