+++ src/pages/apply/components/ApplyQualify.tsx
@@ -6,7 +6,7 @@ import ApplyInterview from '@gdsc/assets/ApplyInterview.svg';
 import ApplyResult from '@gdsc/assets/ApplyResult.svg';
 
 import styled from '@emotion/styled';
-import { ApplyExInterface } from '@gdsc/interface/ApplyInterface';
+import { ApplyExInterface } from '@gdsc/types/ApplyInterface';
 
 const QualifyContainer = styled.div`
   width: 100%;
